Crypto.Hash.CryptoAPI

Description

Cryptohash API exported through crypto-api.

Note: Current version (0.10) of crypto-api suffers a small performance problem. see http://tab.snarc.org/others/benchmark-cryptohash-0.8.html. Hopefully, future versions will fix this.

Documentation

data MD2

data MD4

data MD5

data SHA1

data SHA224

data SHA256

data SHA384

data SHA512

data Skein256_256

data Skein512_512

data RIPEMD160

data Tiger

data Whirlpool

class (Serialize d, Eq d, Ord d) => Hash ctx d | d -> ctx, ctx -> d where

The Hash class is intended as the generic interface targeted by maintainers of Haskell digest implementations. Using this generic interface, higher level functions such as hash and hash' provide a useful API for comsumers of hash implementations.

Any instantiated implementation must handle unaligned data.

Minimum complete definition: outputLength, blockLength, initialCtx, updateCtx, and finalize.

Minimal complete definition

outputLength, blockLength, initialCtx, updateCtx, finalize

Contexts

data CTXMD2

data CTXMD4

data CTXMD5

data CTXRIPEMD160

data CTXSHA1

data CTXSHA224

data CTXSHA256

data CTXSHA384

data CTXSHA512

data CTXSkein256_256

data CTXSkein512_512

data CTXTiger

data CTXWhirlpool